fix: move sudo-helper file to a safe location
Previously, the sudo-helper file was located next to the novus binary.
However, this posed a security threat as the Homebrew folders are not root-owned.
Therefore, even though the sudo-helper file was owned by root, and thus not directly editable,
the attacker could delete the parent directory and replace it with its own sudo-helper file,
containing a malicious code.

This commit fixes it by creating a new root owned directory in /usr/local/novus where the sudo-helper
file will be placed. This folder nor the file cannot be deleted without sudo permissions.